started to split spawner.js into separate files based on old and new spawner code
